--- a/include/tvm/meta_schedule/search_strategy.h
+++ b/include/tvm/meta_schedule/search_strategy.h
@@ -88,6 +88,8 @@ class SearchStrategyNode : public runtime::Object {
 
   /*!
    * \brief Pre-tuning for the search strategy.
+   * \param max_trials The maximum number of trials.
+   * \param num_trials_per_iter The number of trials per iteration.
    * \param design_spaces The design spaces used during tuning process.
    * \param database The database used during tuning process.
    * \param cost_model The cost model used during tuning process.
@@ -95,7 +97,8 @@ class SearchStrategyNode : public runtime::Object {
    *  initialization. Because the search strategy is stateful, we can always call pretuning
    *  and reset the search strategy.
    */
-  virtual void PreTuning(const Array<tir::Schedule>& design_spaces,
+  virtual void PreTuning(int max_trials, int num_trials_per_iter,
+                         const Array<tir::Schedule>& design_spaces,
                          const Optional<Database>& database,
                          const Optional<CostModel>& cost_model) = 0;
 
@@ -143,10 +146,10 @@ class SearchStrategy : public runtime::ObjectRef {
   using FInitializeWithTuneContext = runtime::TypedPackedFunc<void(const TuneContext&)>;
   /*!
    * \brief The function type of `PreTuning` method.
-   * \param design_spaces The design spaces for pre-tuning.
    */
   using FPreTuning = runtime::TypedPackedFunc<void(
-      const Array<tir::Schedule>&, const Optional<Database>&, const Optional<CostModel>&)>;
+      int max_trials, int num_trials_per_iter, const Array<tir::Schedule>&,
+      const Optional<Database>&, const Optional<CostModel>&)>;
   /*! \brief The function type of `PostTuning` method. */
   using FPostTuning = runtime::TypedPackedFunc<void()>;
   /*!
@@ -185,24 +188,15 @@ class SearchStrategy : public runtime::ObjectRef {
 
   /*!
    * \brief Constructor of replay trace search strategy.
-   * \param num_trials_per_iter The number of trials per iteration, i.e., the batch size.
-   * \param max_trials_per_task The total number of trials for trace replaying.
    * \param max_fail_count The max number of failures during trace replaying.
    */
-  TVM_DLL static SearchStrategy ReplayTrace(int num_trials_per_iter, int max_trials_per_task,
-                                            int max_fail_count);
+  TVM_DLL static SearchStrategy ReplayTrace(int max_fail_count);
 
-  /*!
-   * \brief Constructor of replay func search strategy.
-   * \param num_trials_per_iter The number of trials per iteration, i.e., the batch size.
-   * \param max_trials_per_task The total number of trials for func replaying.
-   */
-  TVM_DLL static SearchStrategy ReplayFunc(int num_trials_per_iter, int max_trials_per_task);
+  /*! \brief Constructor of replay func search strategy. */
+  TVM_DLL static SearchStrategy ReplayFunc();
 
   /*!
    * \brief Constructor of evolutionary search strategy.
-   * \param num_trials_per_iter The number of trials per iteration, i.e., the batch size.
-   * \param max_trials_per_task The total number of trials for evolutionary search.
    * \param population_size The initial sample population.
    * \param init_measured_ratio The ratio of measures samples in initial population.
    * \param init_min_unmeasured The minimal size of unmeasured population in the initial sampling.
@@ -211,9 +205,7 @@ class SearchStrategy : public runtime::ObjectRef {
    * \param genetic_max_fail_count The maximum number to try evolving the given trace.
    * \param eps_greedy The ratio to select samples in a greedy fashion via their predicted score.
    */
-  TVM_DLL static SearchStrategy EvolutionarySearch(int num_trials_per_iter,     //
-                                                   int max_trials_per_task,     //
-                                                   int population_size,         //
+  TVM_DLL static SearchStrategy EvolutionarySearch(int population_size,         //
                                                    double init_measured_ratio,  //
                                                    int init_min_unmeasured,     //
                                                    int genetic_num_iters,       //
@@ -257,8 +249,8 @@ class PySearchStrategyNode : public SearchStrategyNode {
   }
 
   void InitializeWithTuneContext(const TuneContext& context) final;
-  void PreTuning(const Array<tir::Schedule>& design_spaces, const Optional<Database>& database,
-                 const Optional<CostModel>& cost_model) final;
+  void PreTuning(int max_trials, int num_trials_per_iter, const Array<tir::Schedule>& design_spaces,
+                 const Optional<Database>& database, const Optional<CostModel>& cost_model) final;
   void PostTuning() final;
   Optional<Array<MeasureCandidate>> GenerateMeasureCandidates() final;
   void NotifyRunnerResults(const Array<MeasureCandidate>& measure_candidates,