--- a/include/tvm/meta_schedule/space_generator.h
+++ b/include/tvm/meta_schedule/space_generator.h
@@ -20,10 +20,14 @@
 #define TVM_META_SCHEDULE_SPACE_GENERATOR_H_
 
 #include <tvm/ir/module.h>
+#include <tvm/meta_schedule/mutator.h>
+#include <tvm/meta_schedule/postproc.h>
+#include <tvm/meta_schedule/schedule_rule.h>
 #include <tvm/node/reflection.h>
 #include <tvm/runtime/container/array.h>
 #include <tvm/runtime/object.h>
 #include <tvm/runtime/packed_func.h>
+#include <tvm/target/target.h>
 #include <tvm/tir/schedule/schedule.h>
 
 namespace tvm {
@@ -71,6 +75,19 @@ class SpaceGenerator;
 */
 class SpaceGeneratorNode : public runtime::Object {
  public:
+  /*! \brief The schedule rules. */
+  Optional<Array<ScheduleRule>> sch_rules;
+  /*! \brief The postprocessors. */
+  Optional<Array<Postproc>> postprocs;
+  /*! \brief The probability of using certain mutator. */
+  Optional<Map<Mutator, FloatImm>> mutator_probs;
+
+  void VisitAttrs(tvm::AttrVisitor* v) {
+    v->Visit("sch_rules", &sch_rules);
+    v->Visit("postprocs", &postprocs);
+    v->Visit("mutator_probs", &mutator_probs);
+  }
+
   /*! \brief Default destructor */
   virtual ~SpaceGeneratorNode() = default;
 
@@ -79,7 +96,7 @@ class SpaceGeneratorNode : public runtime::Object {
    * \param context The tuning context for initialization.
    * \note This method is supposed to be called only once before every other method.
    */
-  virtual void InitializeWithTuneContext(const TuneContext& context) = 0;
+  virtual void InitializeWithTuneContext(const TuneContext& context);
 
   /*!
    * \brief Generate design spaces given a module.
@@ -127,12 +144,17 @@ class SpaceGenerator : public runtime::ObjectRef {
  public:
   /*!
    * \brief Create a design space generator with customized methods on the python-side.
+   * \param sch_rules The schedule rules.
+   * \param postprocs The postprocessors.
+   * \param mutator_probs The probability of using certain mutator.
    * \param f_initialize_with_tune_context The packed function of `InitializeWithTuneContext`.
    * \param f_generate_design_space The packed function of `GenerateDesignSpace`.
    * \param f_clone The packed function of `Clone`.
    * \return The design space generator created.
    */
   TVM_DLL static SpaceGenerator PySpaceGenerator(
+      Optional<Array<ScheduleRule>> sch_rules, Optional<Array<Postproc>> postprocs,
+      Optional<Map<Mutator, FloatImm>> mutator_probs,
       FInitializeWithTuneContext f_initialize_with_tune_context,
       FGenerateDesignSpace f_generate_design_space, FClone f_clone);
   /*!
@@ -141,19 +163,39 @@ class SpaceGenerator : public runtime::ObjectRef {
    * 1) void(Schedule)
    * 2) Schedule(Schedule)
    * 3) Array<Schedule>(Schedule)
+   * \param sch_rules The schedule rules.
+   * \param postprocs The postprocessors.
+   * \param mutator_probs The probability of using certain mutator.
    */
-  TVM_DLL static SpaceGenerator ScheduleFn(PackedFunc schedule_fn);
+  TVM_DLL static SpaceGenerator ScheduleFn(PackedFunc schedule_fn,
+                                           Optional<Array<ScheduleRule>> sch_rules,
+                                           Optional<Array<Postproc>> postprocs,
+                                           Optional<Map<Mutator, FloatImm>> mutator_probs);
   /*!
    * \brief Create a design space generator that is union of multiple design space generators.
    * \param space_generators An array of design space generators to be unioned.
+   * \param sch_rules The schedule rules.
+   * \param postprocs The postprocessors.
+   * \param mutator_probs The probability of using certain mutator.
    * \return The design space generator created.
    */
-  TVM_DLL static SpaceGenerator SpaceGeneratorUnion(Array<SpaceGenerator, void> space_generators);
+  TVM_DLL static SpaceGenerator SpaceGeneratorUnion(Array<SpaceGenerator, void> space_generators,
+                                                    Optional<Array<ScheduleRule>> sch_rules,
+                                                    Optional<Array<Postproc>> postprocs,
+                                                    Optional<Map<Mutator, FloatImm>> mutator_probs);
   /*!
    * \brief Create a design space generator that generates design spaces by applying schedule
-   * rules to blocks in post-DFS order. \return The design space generator created.
+   * rules to blocks in post-DFS order.
+   * \param f_block_filter The filter function to filter blocks to be applied with schedule rules.
+   * \param sch_rules The schedule rules.
+   * \param postprocs The postprocessors.
+   * \param mutator_probs The probability of using certain mutator.
+   * \return The design space generator created.
    */
-  TVM_DLL static SpaceGenerator PostOrderApply(runtime::PackedFunc f_block_filter = nullptr);
+  TVM_DLL static SpaceGenerator PostOrderApply(runtime::PackedFunc f_block_filter,
+                                               Optional<Array<ScheduleRule>> sch_rules,
+                                               Optional<Array<Postproc>> postprocs,
+                                               Optional<Map<Mutator, FloatImm>> mutator_probs);
   TVM_DEFINE_MUTABLE_NOTNULLABLE_OBJECT_REF_METHODS(SpaceGenerator, ObjectRef, SpaceGeneratorNode);
 };
 
@@ -171,6 +213,7 @@ class PySpaceGeneratorNode : public SpaceGeneratorNode {
   FClone f_clone;
 
   void VisitAttrs(tvm::AttrVisitor* v) {
+    SpaceGeneratorNode::VisitAttrs(v);
     // `f_initialize_with_tune_context` is not visited
     // `f_generate_design_space` is not visited
     // `f_clone` is not visited