--- a/python/tvm/meta_schedule/cost_model/random_model.py
+++ b/python/tvm/meta_schedule/cost_model/random_model.py
@@ -19,12 +19,11 @@ Random cost model
 """
 from typing import List, Optional, Tuple, Union
 
-from tvm.meta_schedule.utils import derived_object  # type: ignore
-
 from ..cost_model import PyCostModel
 from ..runner import RunnerResult
 from ..search_strategy import MeasureCandidate
 from ..tune_context import TuneContext
+from ..utils import derived_object  # type: ignore
 
 
 @derived_object