@tqchen 2022-09-25
great to see clear separation between tuning and apply. A positive step towards modular automation