@junrushao 2022-10-06
Hey @masahi, I added `executor` parameter to `extract_tasks` and `compile_relay`, which controls the default value of `relay.FuseOps.link_params` in pass configuration. It's quite confusing to me that `executor` is lifted out of pass config and somehow control the compilation process in a half-functioning way (only works for `GraphExecutor`), and am not sure if I'm using that correctly, so please feel free to suggest what the best way is :-)

On the other hand, as a high-level API, I would prefer not to tweak `tune_relay` adding more parameters to it, given we wanted to give a cleaner interface to introductory level users. Instead, advance users could always use `extract_tasks` + `tune_tasks` + `compile_relay` to get fine-grained control over the tuning process