@masahi 2022-10-06
> On the other hand, as a high-level API, I would prefer not to tweak tune_relay adding more parameters to it, given we wanted to give a cleaner interface to introductory level users. Instead, advance users could always use extract_tasks + tune_tasks + compile_relay to get fine-grained control over the tuning process

Yes, I agree with this. A part of the reason I didn't want to change `extract_task` API before was that the vast majority of users don't need to care about `executor` stuff. For Hexagon users, we can add a wrapper API in `contrib/hexagon/meta_schedule` to simplify the usage. We already require using Hexagon-specific builder / runner, so the wrapper API can hide such details too.