@masahi 2022-10-07
@junrushao I made one comment but otherwise Hexagon change looks good to me. It didn't occur to me before that we can do `mod = mod.with_attr(...)` from the user script to avoid threading `executor` through task extraction and `tune_relay` etc.